+++ release/types/column-mode.type.js
@@ -1,8 +1,0 @@
-"use strict";
-(function (ColumnMode) {
-    ColumnMode[ColumnMode["standard"] = 'standard'] = "standard";
-    ColumnMode[ColumnMode["flex"] = 'flex'] = "flex";
-    ColumnMode[ColumnMode["force"] = 'force'] = "force";
-})(exports.ColumnMode || (exports.ColumnMode = {}));
-var ColumnMode = exports.ColumnMode;
-//# sourceMappingURL=column-mode.type.js.map