+++ release/types/selection.type.js
@@ -1,9 +1,0 @@
-"use strict";
-(function (SelectionType) {
-    SelectionType[SelectionType["single"] = 'single'] = "single";
-    SelectionType[SelectionType["multi"] = 'multi'] = "multi";
-    SelectionType[SelectionType["multiShift"] = 'multiShift'] = "multiShift";
-    SelectionType[SelectionType["cell"] = 'cell'] = "cell";
-})(exports.SelectionType || (exports.SelectionType = {}));
-var SelectionType = exports.SelectionType;
-//# sourceMappingURL=selection.type.js.map